rename odf.py to odf_parser.py
